Skip logging if device profile is not initialized
Since recentsAnimation and animationCancel come on the event queue, we may skip
recentsAnimation and only process animationCancel if they happen too close to
each other, before the first event is scheduled on eventQueue.
As a result deviceProfile is never initialized.
Prevent crash by skipping logging in this case, we shouldn't be logging it anyway

Bug: 80515833
Change-Id: Ife1d6009bcdc34af6c281d848665145cd235fa40