Track shelf rather than app window
Change swipe length to be the height of the shelf rather than the distance
to the bottom of the task in recents. The shelf already uses this distance
to determine its starting offset, and the window uses it to determine how
fast to move. So by setting the distance to the height of the shelf, we
effectively make the window move slightly faster and the shelf move directly
under your finger.

Bug: 109817942
Change-Id: I9b20bbed58cccc6f74aba36000a6ca272f9d554d