Swipe up overshoot always plays
Instead of using an OvershootInterpolator, we adjust the end progress to > 1 and
add a second interpolator to settle back to 1. That way, even if the animation runs
starting very late, e.g. 1.0, it still has room to overshoot. We use this same
OvershootParams class to calculate an overshoot for a blocked long fling as well.

Bug: 109709720
Change-Id: I43152237e4350f93e7c462c22e68d09d05c1dd57