Improve adaptive icon generation
The parameters in utitlites#removeColor probably still need a lot of tuning and we might need to add more conditions. A very small number of icons (like the Analytics or google dialer ones) get practically entirely eaten up, this has to be addressed in some way, potentially by limiting the amount of pixels to replace before falling back to the full bitmap and a muted background.