Fix bug where icon appears large until transluscent activity finishes.
It appears large because the pressed state of the view increases the scale,
and we want to leave the pressed state until the activity is launched.

With the remote app transitions, we now have a callback for when the activity
is done launching so we then cancel the "stayPressed" state.

Bug: 111742067
Change-Id: I25725e908b12f4dd1256294a26232b289db2d057