Use default animation when launching a TaskView that's off screen
Now that pressing back launches the previously running task, it's
possible to get TaskView#launch() when the TaskView is not on the
screen (i.e. user scrolled before pressing back). In this case, we
shouldn't use the launch-from-recents animation, as that animation
assumes the task view is the current page or an adjacent one.

Change-Id: I7bec91c9437954061affdedc8cd9e48820ce7357