tapl should be java_library
Bug: 113683301

Tests failures that we are trying to fix.
https://sponge.corp.google.com/invocations?searchFor=target%3Aandroid.platform.systemui.tests.jank%20label%3Ab%2Fgit_master

When the build rule is android_library, AndroidManifest.xml is pulled into
the the final apk that is being built. Hence, the platform jank and platform test
AndroidManifest.xml is extended with the one we have for Launcher3.
This causes the tests to fail due to existence of disambiguation screen.
We do not want this behavior.

Change-Id: I1a551abf37bb75b4efa8a5e5324ab1b4d8a3eb13