Creating a new dismiss animation while one is pending is now valid
It used to be an invalid state to create a new dismiss animation while
one was pending, because the only way to do it was to swipe up. However,
now there are a number of places we create the dismiss animation, e.g.:
- When entering split screen
- Clear all
- Uninstalling an app

So now, instead of calling this an illegal state, we simply cancel the
previous pending dismiss if it exists.

Bug: 118880699
Bug: 118400078
Change-Id: I4d073f7e9d50e832ba4ae55391530040a0fac50c