[DO NOT MERGE] Fix clipping issue on adaptive icons
Bug: 116527322

More context: unlike previous issue that resulted in clipping on adaptive
icon due to launcher drawing issue (drawing outside the bounds, etc).
This issue is happening on b4 because when canvas is scaled after
the circle mask is digitized, there is a optical illusion that one line
of pixel seems missing. The issue is non observed if scaling is done
before setting the view bounds.

Change-Id: I7099075c88e77776c9cc4f23d79152293489cca2