Remove notification dot if app is disabled
For instance, when turning off work profile, all work profile apps are
disabled and notifications are removed. Previously, we weren't updating
the dots in this case because we ignore unsupported ItemInfos (e.g.
disabled items in this case).

Note that this was only user-visible for Folders, because regular
shortcuts were updated when the disabled state changes.

Bug: 111791593
Change-Id: I3c49e6e3bf18753a868eb993b29c77abbefd3466