Fixing flakiness of ShortcutsToHomeTest
More precisely, an experiment on whether this will help. The theory is
that the current implementation where we long-press an icon for a fixed
period of time sometimes doesn't work because Launcher might start
measuring the press time with some delay.

Now we keep pressed the icon until the menu appears.

Change-Id: I1cead505726a998316a73d81b10160517f17f08f
Tests:ShortcutsToHomeTest.testDragIcon