Fix RecentTasksList change id incrementing
The change id in RecentTasksList would never actually increment as we
never register it as a TaskStackListener. As a result, we always execute
the result callback immediately when getting a task and the task list is
always valid.

In addition, when the list IS up to date, we still fetch the list again
in the background when we should instead just return early.

This CL addresses both of these issues.

Test: Added logs to see that task loading wasn't happening when lists
are the same

Change-Id: I228f8d7cd3cb22ef88d71e1bee40708c2bcf26d6