Send onGestureStarted() when continuing a quick switch gesture
When continuing the gesture, we set mPassedTouchSlop = true in order to
make the second gesture responsive/fluid. However, we only call
notifyGestureStarted() when passing that slop, meaning we don't call it
in this case. Now we call it immediately on touch down when continuing
the previous gesture.

This cleans up a lot of inconsistent state issues, such as not creating
the launcher animation controller as part of the second quick switch
gesture.

Bug: 111926330
Change-Id: I03c3ed9a8d364fbfdc2d04bafb58a63bf61e2f63