Don't shrink overview bitmap for chips
Chips in overview will be shown on top of the bitmap, not in
between the bitmap and the QSB, so we no longer need to shrink
the bitmap to make room for chips. This means that the appearance
of the recents view will be consistent between chips being
enabled/disabled.

Change-Id: I1856ccb8babd388ef6c8e6a1a5defcbdd0278a60