Don't layout empty tasks in layout anim if we have the actual content.
When we load the task content (icons, snapshots, etc), oftentimes, we
are still in the middle of the layout animation where items are fading
in from bottom to top. At this point, we start a second content-fill
animation from bottom to top that fades from empty => filled after the
first animation.  However, we can improve this so that empty views
aren't shown for a split second for some of the later laid out views.

If the first animation has not finished and we would animate a content
change from empty => filled but the view is currently still not visible,
we should just complete the content transition on the spot so that the
layout animation lays out the filled version instead.

Bug: 131339235
Test: Go to recents, observe lay out animation does not continue to lay
out the empty view first

Change-Id: I936a81e6cf8b3552cdee90c183fc841f50ec9ea8