Fix RTL bug when swiping up to home.
When you switch to RTL when an app is open, we try to use the BubbleTextView
that is no longer attached on the animation back to home. This causes the
item location to return the wrong values (since now the view doesn't have
valid padding/width we can use). In this case, we'll use the fallback
animation.

Bug: 131242470
Change-Id: I0ce27ee10bd6402cc6b2065813b034fc04592622