Allow touches on launcher while animating to home
Previously we only proxied if the end target was recents (now we do it
for home as well) and only proxied to the controllers, not to all views.
We still only proxy to controllers for touches in the nav bar, but
otherwise we dispatch touches through normal view hierarchy.

Tests while animating home:
- Swipe down for notifications
- Swipe up for all apps
- Long press icon or wallpaper for shortcuts
- Tap a folder to open it
- Tap an icon to launch it (currently no animation plays, looking into
  it separately)

Bug: 131775822
Change-Id: I705b7a657b6d17bc0e24ed4fb44ecfb0f319e942